Major refactor. Use dynamic discovery model.
The project was based on defining in code each option and command
available by the REST API.

This generates a lot of boilerplate code and makes it harder to use with
custom APIs deployed at the server.

By using a dynamic discovery model, the CLI will recognize custom APIs
deployed at the server, as long as they are listed in the API description
response.

We lose some labels for each field, but neither the API gives any
description about them.

This commits creates a new module called 'routes' that uses a description
JSON from the API to build up every option and command available. This
module deprecates the modules 'comments', 'media', 'pages', 'posts' and
'taxonomies'.

It is a fast-coded module, may need some refactorings to improve
readability.

Didn't fully delete other modules because I'm not sure this is the correct
way to go. Will give it some time.