Use argument description from schema definition.
If the name of the argument of a command matched the name used as a
top-level schema definition property of the same endpoint, use the
description available for the schema property as a description for the
argument.

This only affects the output of command 'info' and not the help documentation
generated by '--help'. Why? Because the same argument name may have
different descriptions in each schema.