Cache site URL from update command.
Remove the necessity of setting a site to use in every command.

Save the entire description of the API in `update` command instead of
routes only. When loading the file, set the site's URL and base URL from
the description.

User still may override the site to use, as the option `--site` is still
available.